feat(internal): 新增训练任务管理功能
- 新增 TrainingTaskManager 接口，用于管理训练任务
- 实现 TrainingTaskInfo 结构体，用于存储训练任务信息
- 新增相关事件消息结构体，用于训练任务的请求和响应
- 重构部分代码，以支持新的训练任务管理功能

+++ internal/data/gpu_nvidia.go
@@ -66,7 +66,7 @@ func (m *NvidiaGpuManager) getNvidiaGpuMetrics() []event.GpuInfo {
 			}
 
 			gpuInfo := event.GpuInfo{
-				Index:       index,
+				Index:       int32(index),
 				Name:        metrics[1],
 				CudaUsage:   float64(int(cudaUsage*100)) / 100.0, // 保留两位小数
 				MemoryTotal: memoryTotal,
